Split telemetry tests
